add ability to log out and remove unused file
